requirements.txt
openai>=0.29.0
python-dotenv>=1.0.0
yfinance>=0.2.0
pandas>=2.0.0
python-telegram-bot>=20.0
jinja2>=3.0.0
pytz>=2023.3
